+++ IJPLib_Test/IJPLib_Test/IJPLib_Test.csproj
@@ -66,6 +66,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils.cs" />
+    <Compile Include="XmlToMsg.cs" />
     <EmbeddedResource Include="IJPTest.resx">
       <DependentUpon>IJPTest.cs</DependentUpon>
     </EmbeddedResource>
